copilot/requirements.txt
openai==1.54.3
fastapi==0.115.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
numpy==1.26.4
pgvector==0.3.6
pydantic==2.9.2
llama-index-core==0.11.22
llama-cloud==0.1.4
uvicorn==0.32.0
websockets==14.0
llama-index-vector-stores-postgres==0.2.6
